+++ src/main.rs
@@ -1,5 +1,9 @@
+mod notify;
+mod pomodoro;
+
 use crossbeam::{channel::bounded, select};
-use rust_pomodoro::Pomodoro;
+use notify::notify;
+use pomodoro::Pomodoro;
 use std::{process, thread};
 
 fn main() {
@@ -17,18 +21,18 @@ fn main() {
         select! {
             recv(&is_start_receiver) -> p => {
                 if let Ok(_) = p {
-                    rust_pomodoro::notify(String::from("Starting"));
+                   notify(String::from("Starting"));
                 }
             }
             recv(&is_done_receiver) -> p => {
                 if let Ok(_) = p {
-                    rust_pomodoro::notify(String::from("Finish"));
+                    notify(String::from("Finish"));
                 }
                 process::exit(1);
             }
             recv(&is_break_receiver) -> p => {
                 if let Ok(_) = p {
-                    rust_pomodoro::notify(String::from("Break"));
+                    notify(String::from("Break"));
                 }
             }
         }
